+++ web/documentserver-example/csharp/Users.cs
@@ -78,7 +78,8 @@ namespace OnlineEditorsExample
             "Can't protect file",
             "View file without collaboration",
             "Can’t submit forms",
-            "Can’t refresh outdated file"
+            "Can’t refresh outdated file",
+            "Tour of tips when opening a document"
         };
 
         private static List<User> users = new List<User>() {